telegram-bot/bot/requirements.txt
aiogram==3.18.0
apscheduler==3.10.1
pytz==2023.3.post1
loguru==0.7.2
python-dotenv==1.0.0